Add support for value maps
In order to simplify individual fields, this is done in the base layer so they can all inherit it. This also means that all fields must make sure to call super() during encode and decode in order to support maps. Failing to call super() will still work for most cases, but map support won't work.