Pull out lazy attribute loading
The .load() method in marshal/pickle/json loads everything immediately, and that's the most straightforward behavior to implement. The lazy aspect may prove useful, but it's too early to try to work that in while everything else is still in flux. Once it's thought out more, I'll work to bring it back.