Add StructureTuple support
This is a neat little hack that allows a structure to instantiate itself as a named tuple instead of a full object. For objects that just need to be read and don't have any custom behavior, this can provide a significant memory savings, particularly if it's a structure that occurs frequently in a file.